[SyncLogging] Extract logging out of Sync
Single Responsibility principle - Sync knows nothing about how it
logs, it just delegates to SyncLogging.
